CLEAN: move json formatter to a separate file
